Minor changes in Docs and CMakeLists
Move png files to doc-files in order to let all documentation files in a
single directory. It was necessary to fix README files to adjust the
paths of png files.

Change the way the installation is done in Arduino target.
Now you have to pass the port to build-arduino-micro.sh script.
The doc was adjusted according to that.